Fixes all unit tests.
