+++ Aztec/Classes/NSAttributedString/Conversions/AttributedStringSerializer.swift
@@ -119,11 +119,10 @@ class AttributedStringSerializer {
     ///
     fileprivate func serialize(_ element: ElementNode, inheriting attributes: [NSAttributedStringKey: Any]) -> NSAttributedString {
 
-        let childAttributes = self.attributes(for: element, inheriting: attributes)
         let content = NSMutableAttributedString()
 
         let converter = self.converter(for: element)
-        let convertedString = converter.convert(element, inheriting: childAttributes, childrenSerializer: childrenSerializer)
+        let convertedString = converter.convert(element, inheriting: attributes, childrenSerializer: childrenSerializer)
         
         content.append(convertedString)
         
@@ -144,27 +143,6 @@ class AttributedStringSerializer {
 
         return NSAttributedString(attributedString: stringWithSeparator)
     }
-
-    // MARK: - Built-in formatter instances
-
-    lazy var blockquoteFormatter = BlockquoteFormatter()
-    lazy var boldFormatter = BoldFormatter()
-    lazy var divFormatter = HTMLDivFormatter()
-    lazy var h1Formatter = HeaderFormatter(headerLevel: .h1)
-    lazy var h2Formatter = HeaderFormatter(headerLevel: .h2)
-    lazy var h3Formatter = HeaderFormatter(headerLevel: .h3)
-    lazy var h4Formatter = HeaderFormatter(headerLevel: .h4)
-    lazy var h5Formatter = HeaderFormatter(headerLevel: .h5)
-    lazy var h6Formatter = HeaderFormatter(headerLevel: .h6)
-    lazy var italicFormatter = ItalicFormatter()
-    lazy var linkFormatter = LinkFormatter()
-    lazy var orderedListFormatter = TextListFormatter(style: .ordered, increaseDepth: true)
-    lazy var paragraphFormatter = HTMLParagraphFormatter()
-    lazy var preFormatter = PreFormatter()
-    lazy var strikethroughFormatter = StrikethroughFormatter()
-    lazy var underlineFormatter = UnderlineFormatter()
-    lazy var unorderedListFormatter = TextListFormatter(style: .unordered, increaseDepth: true)
-    lazy var codeFormatter = CodeFormatter()
 
     // MARK: - Built-in element converter instances
 
@@ -194,29 +172,6 @@ class AttributedStringSerializer {
         return [:]
     }()
 
-    public lazy var elementFormattersMap: [Element: AttributeFormatter] = {
-        return [
-            .blockquote: self.blockquoteFormatter,
-            .div: self.divFormatter,
-            .ol: self.orderedListFormatter,
-            .ul: self.unorderedListFormatter,
-            .strong: self.boldFormatter,
-            .em: self.italicFormatter,
-            .u: self.underlineFormatter,
-            .del: self.strikethroughFormatter,
-            .a: self.linkFormatter,
-            .h1: self.h1Formatter,
-            .h2: self.h2Formatter,
-            .h3: self.h3Formatter,
-            .h4: self.h4Formatter,
-            .h5: self.h5Formatter,
-            .h6: self.h6Formatter,
-            .p: self.paragraphFormatter,
-            .pre: self.preFormatter,
-            .code: self.codeFormatter
-        ]
-    }()
-
     let attributesToFormattersMap: [StandardHTMLAttribute: AttributeFormatter] = [:]
 
     public let styleToFormattersMap: [String: (AttributeFormatter, (String)->Any?)] = [
@@ -238,134 +193,6 @@ class AttributedStringSerializer {
         }
         return stylesDictionary
     }
-}
-
-private extension AttributedStringSerializer {
-
-    // MARK: - NSAttributedString attribute generation
-
-    /// Calculates the attributes for the specified node.  Returns a dictionary including inherited
-    /// attributes.
-    ///
-    /// - Parameters:
-    ///     - element: the node to get the information from.
-    ///     - inheritedAttributes: the inherited attributes from parent nodes.
-    ///
-    /// - Returns: an attributes dictionary, for use in an NSAttributedString.
-    ///
-    func attributes(for element: ElementNode, inheriting inheritedAttributes: [NSAttributedStringKey: Any]) -> [NSAttributedStringKey: Any] {
-
-        guard !(element is RootNode) else {
-            return inheritedAttributes
-        }
-
-        let elementRepresentation = HTMLElementRepresentation(element)
-        let representation = HTMLRepresentation(for: .element(elementRepresentation))
-        var finalAttributes = inheritedAttributes
-
-        if let elementFormatter = formatter(for: element) {
-            finalAttributes = elementFormatter.apply(to: finalAttributes, andStore: representation)
-        } else if element.type == .li || hasSpecializedConverter(for: element) {
-            finalAttributes = inheritedAttributes
-        } else {
-            finalAttributes = self.attributes(storing: elementRepresentation, in: finalAttributes)
-        }
-
-        finalAttributes = self.attributes(for: element.attributes, inheriting: finalAttributes)
-        
-        return finalAttributes
-    }
-
-    /// Calculates the attributes for the specified HTML attributes.  Returns a dictionary
-    /// including the inherited attributes.
-    ///
-    /// - Parameters:
-    ///     - htmlAttributes: the HTML attributes to calculate the string attributes from.
-    ///     - inheritedAttributes: the attributes that will be inherited.
-    ///
-    /// - Returns: an attributes dictionary, for use in an NSAttributedString.
-    ///
-    private func attributes(for htmlAttributes: [Attribute], inheriting inheritedAttributes: [NSAttributedStringKey: Any]) -> [NSAttributedStringKey: Any] {
-
-        let finalAttributes = htmlAttributes.reduce(inheritedAttributes) { (previousAttributes, htmlAttribute) -> [NSAttributedStringKey: Any] in
-            return attributes(for: htmlAttribute, inheriting: previousAttributes)
-        }
-
-        return finalAttributes
-    }
-
-
-    /// Calculates the attributes for the specified HTML attribute.  Returns a dictionary
-    /// including inherited attributes.
-    ///
-    /// - Parameters:
-    ///     - attribute: the attribute to calculate the string attributes from.
-    ///     - inheritedAttributes: the attributes that will be inherited.
-    ///
-    /// - Returns: an attributes dictionary, for use in an NSAttributedString.
-    ///
-    private func attributes(for attribute: Attribute, inheriting inheritedAttributes: [NSAttributedStringKey: Any]) -> [NSAttributedStringKey: Any] {
-
-        let attributes: [NSAttributedStringKey: Any]
-
-        if let attributeFormatter = formatter(for: attribute) {
-            let attributeHTMLRepresentation = HTMLRepresentation(for: .attribute(attribute))
-
-            attributes = attributeFormatter.apply(to: inheritedAttributes, andStore: attributeHTMLRepresentation)
-        } else {
-            attributes = inheritedAttributes
-        }
-        
-        return attributes
-    }
-
-
-    /// Stores the specified HTMLElementRepresentation in a collection of NSAttributedString Attributes.
-    ///
-    /// - Parameters:
-    ///     - representation: Instance of HTMLElementRepresentation to be stored.
-    ///     - attributes: Attributes where we should store the HTML Representation.
-    ///
-    /// - Returns: A collection of NSAttributedString Attributes, including the specified HTMLElementRepresentation.
-    ///
-    private func attributes(storing representation: HTMLElementRepresentation, in attributes: [NSAttributedStringKey: Any]) -> [NSAttributedStringKey: Any] {
-        let unsupportedHTML = attributes[.unsupportedHtml] as? UnsupportedHTML
-        var representations = unsupportedHTML?.representations ?? []
-        representations.append(representation)
-
-        // Note:
-        // We'll *ALWAYS* store a copy of the UnsupportedHTML instance. Reason is: reusing the old instance
-        // would mean affecting a range that may fall beyond what we expected!
-        //
-        var updated = attributes
-        updated[.unsupportedHtml] = UnsupportedHTML(representations: representations)
-
-        return updated
-    }
-}
-
-extension AttributedStringSerializer {
-
-    // MARK: - Formatters
-
-    func formatter(for attribute: Attribute) -> AttributeFormatter? {
-        // TODO: implement attribute representation formatters
-        //
-        return nil
-    }
-
-    func formatter(for element: ElementNode) -> AttributeFormatter? {
-        let equivalentNames = element.type.equivalentNames
-
-        for (key, formatter) in elementFormattersMap {
-            if equivalentNames.contains(key.rawValue) {
-                return formatter
-            }
-        }
-
-        return nil
-    }
-
 }
 
 private extension AttributedStringSerializer {
