+++ Aztec/Classes/NSAttributedString/Conversions/StringToNodeConverters/Base/StringToNodeConverter.swift
@@ -1,0 +1,5 @@
+import Foundation
+
+public protocol StringToNodeConverter {
+    func convert(_ attrString: NSAttributedString) -> Node?
+}
